fix: split apart the calendar
