+++ app/components/ui/Calendar/constants.ts
@@ -1,0 +1,23 @@
+// Color palette for categories
+export const CATEGORY_COLORS = [
+  "#E3F2FD", // Light Blue
+  "#E8F5E8", // Light Green
+  "#F3E5F5", // Light Purple
+  "#FFF3E0", // Light Orange
+  "#FCE4EC", // Light Pink
+  "#FFFDE7", // Light Yellow
+  "#E0F7FA", // Light Cyan
+  "#F0F4FF", // Light Periwinkle
+  "#F5F5F5", // Light Gray
+];
+
+// Map of day abbreviations to JavaScript day numbers (0 = Sunday, 6 = Saturday)
+export const DAY_MAP: { [key: string]: number } = {
+  Su: 0,
+  M: 1,
+  T: 2,
+  W: 3,
+  Th: 4,
+  F: 5,
+  Sa: 6,
+};
