+++ app/components/ui/Calendar/types.ts
@@ -1,0 +1,45 @@
+import { Event } from "react-big-calendar";
+import { SFGovEvent } from "hooks/SFGovAPI";
+
+export interface CalendarEvent extends Event {
+  id: string;
+  pageLink: string;
+  description: string;
+  location?: string;
+  originalEvent: SFGovEvent;
+}
+
+export interface EventCalendarProps {
+  onEventSelect?: (event: CalendarEvent) => void;
+}
+
+export interface CategoryFilter {
+  category: string;
+  enabled: boolean;
+  color: string;
+}
+
+export interface CategoryFiltersProps {
+  availableCategories: string[];
+  categoryFilters: CategoryFilter[];
+  onToggleCategory: (category: string) => void;
+}
+
+export interface MobileAgendaProps {
+  currentDate: Date;
+  groupedEvents: { [timeKey: string]: CalendarEvent[] };
+  onNavigatePrevious: () => void;
+  onNavigateNext: () => void;
+  onEventSelect: (event: CalendarEvent) => void;
+  formatDateHeader: (date: Date) => string;
+}
+
+export interface EventSlideoutProps {
+  isOpen: boolean;
+  onClose: () => void;
+  selectedEvent: CalendarEvent | null;
+  dayEvents: CalendarEvent[];
+  selectedDate: Date | null;
+  categoryColorMap: Map<string, string>;
+  onEventSelect: (event: CalendarEvent) => void;
+}
