+++ app/pages/HomePage/HomePage.tsx
@@ -2,9 +2,8 @@ import React from "react";
 import Hero from "components/ui/Hero/Hero";
 import { CategorySection } from "components/ui/Section/CategorySection";
 import { useHomepageData, useHomePageEventsData } from "hooks/StrapiAPI";
-import { useAllSFGovEvents } from "hooks/SFGovAPI";
 import { Homepage, StrapiDatum } from "models/Strapi";
-import { EventCalendar } from "components/ui/Calendar/EventCalendar";
+import { EventCalendar } from "components/ui/Calendar";
 import { HomePageSection } from "pages/HomePage/components/Section";
 import { TwoColumnContentSection } from "components/ui/TwoColumnContentSection/TwoColumnContentSection";
 import { EventCardSection } from "components/ui/Cards/EventCardSection";
@@ -12,7 +11,6 @@ import { EventCardSection } from "components/ui/Cards/EventCardSection";
 export const HomePage = () => {
   const { data: homepageData, isLoading: homepageDataIsLoading } =
     useHomepageData();
-  const { data: eventsData, isLoading: eventsAreLoading } = useAllSFGovEvents();
 
   const { data: featuredEventsData, isLoading: featuredEventsAreLoading } =
     useHomePageEventsData();
@@ -21,11 +19,7 @@ export const HomePage = () => {
 
   const homePageDataAttrs = homepageDataRes?.attributes;
 
-  if (
-    homepageDataIsLoading ||
-    eventsAreLoading ||
-    (featuredEventsAreLoading && featuredEventsData)
-  ) {
+  if (homepageDataIsLoading) {
     return null;
   }
 
@@ -48,23 +42,23 @@ export const HomePage = () => {
         <CategorySection />
       </span>
 
-      {eventsData && (
-        <span id="featured-events">
-          <HomePageSection
-            title={"Events calendar"}
-            description={""}
-            backgroundColor={"primary"}
-          >
-            <EventCalendar events={eventsData} />
-          </HomePageSection>
-          <HomePageSection
-            title={"Featured resources"}
-            description={""}
-            backgroundColor={"tertiary"}
-          >
-            <EventCardSection events={featuredEventsData ?? []} />
-          </HomePageSection>
-        </span>
+      <span id="featured-events">
+        <HomePageSection
+          title={"Events calendar"}
+          description={""}
+          backgroundColor={"primary"}
+        >
+          <EventCalendar />
+        </HomePageSection>
+      </span>
+      {!featuredEventsAreLoading && (
+        <HomePageSection
+          title={"Featured resources"}
+          description={""}
+          backgroundColor={"tertiary"}
+        >
+          <EventCardSection events={featuredEventsData ?? []} />
+        </HomePageSection>
       )}
 
       {two_column_content_block?.map((content) => (
